Use size from datastore and check to be not more than provided size
We must have size of blob to download parts of it correctly. But seems
current logic has no checks for size provided from configuration. In
this case we can hit the problems when we have miss-configured size, so
lets get size from datastore in all cases and check for mismatch.

Signed-off-by: Petr Fedchenkov <giggsoff@gmail.com>