http: minimize timeout by propagating error
If eve looses connection while downloading an app image, it
first has to timeout to retry downloading with another network
interface.
The reason is that the error from io.Copy (copying from the network
connection) is not propagating the error to the caller that decides
to cancel this connection and retry on a different interface.

This change propagates the error to the downloader goroutine and this
one then cancels the connection.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>