nettrace: skip over interfaces where pcap failed to establish
Some network interfaces selected for packet capture may not be in a
state suitable for packet capture. However, it seems reasonable to
simply report this with error logs and skip over such interfaces rather
than to completely fail to create a traced HTTP client.
In other words, for nettrace we setup pcap in best-effort mode.

Signed-off-by: Milan Lenco <milan@zededa.com>