zedUpload/timeoutreader: fix timer leak which causes OOM
Patch fixes very subtle memory leak which is very well explained here:
https://arangodb.com/2020/09/a-story-of-a-memory-leak-in-go-how-to-properly-use-time-after/

The problem is easy to reproduce: deploy Vm application with the huge
image (~10Gb) and monitor the zedbox process on RSS value while image
is downloaded: RSS reaches ~800Mb and zedbox is killed by the OOM.

Thanks to recently introduced go tracing in pillar (the `eve http-debug`
feature), the memory leak was easy to catch.

Kudos to @rene and @christoph-zededa for helping to reproduce, catch
and debug the issue.

Likely new patches with similar fixes will follow, because similar
code pattern is used in many places.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>