nettrace: add support for local stub resolvers
Added support for operating systems with a local stub resolver
(e.g. systemd-resolved) that listens on a loopback IP.
When sending DNS requests in such environments, the HTTP client must not
force a caller-specified source IP. Both the source and destination IPs
of the DNS request must be loopback addresses for local stub connectivity
to work correctly.

In EVE we do not use a local stub resolver, but this patch is needed
to support running unit tests with nettrace on hosts with different
OS configurations.

Signed-off-by: Milan Lenco <milan@zededa.com>