Update conntrack library to include patch for IPv6
Library github.com/ti-mo/conntrack was patched to properly handle
conntrack retrieval for IPv6 entries.

This is the PR with the fix:
https://github.com/ti-mo/conntrack/pull/46

We use this library inside nettrace to capture information provided by
conntrack, such as the number of packets/bytes seen in both directions,
state of the connection as observed by conntrack, etc.

Signed-off-by: Milan Lenco <milan@zededa.com>