Added support for S3-compatible object stores
This commit adds the support for S3 compatible datastores other than S3 per se (e.g., MinIO).
The main problem that we had is that under the hood AWS SDK v2 for Go sends S3 calls by
consulting a global endpoint resolver and then building a URL like:
https://{bucket}.s3.{region}.amazonaws.com/{key}. When we point to an S3-compatible object
store we need to override the endpoint so that every request is sent to our server URL instead of
s3.amazonaws.com. So, we set the BaseEndpoint to avoid consulting the global resolver, and have
DNS-validation errors about "Invalid region."

Signed-off-by: Ioannis Sfakianakis <jsfakas@gmail.com>